Add support for alternative directorynames of 'persistence' and 'liveslak'.
The persistence directory on USB is called 'persistence' by default but can
now be re-defined when running 'make_slackware_live.sh' to create the ISO.
Additionally it can be changed dynamically on boot of the Live OS by adding
a 'persistence=<dirname>' parameter to the boot.

The Live OS main directory is called 'liveslak' by default but can
now be re-defined when running 'make_slackware_live.sh' to create the ISO.
Additionally it can be changed dynamically on boot of the Live OS by adding
a 'livemain=<dirname>' parameter to the boot.