Much enhanced iso2usb script.
Two new parameters '--force' and '--persistence'. Read "usb2iso.sh --help".
Size of the EFI partition was reduced from 200 to 100 MB.
The wait-for-root time in the initrd.img file is changed for both BIOS
and UEFI boot; this should make boot work out of the box for most computers.
More robustness was added in handling race conditions.
Only make the USB stick UEFI-bootable if the ISO file is capable of the same.