Do not add Nvidia proprietary drivers to the Full Slackware ISO.
The customized ISOs (Slackware plus 3rd party software) can include the
proprietary drivers, but the Full Slackware ISO is meant as a demo for
official Slackware and therefore should not include 3rd party packages.

If you respin the SLACKWARE ISO, it is of course still possible to add the
binary Nvidia blob yourself by copying the graphics/*nvidia*.szx file to
the 'optional' directory yourself. Nothing else is needed, it will 'just' work.