Get rid of udev errors about obex-check-device on boot.
The mkinitrd script adds /lib/udev/rules.d/60-openobex.rules to the initrd
image, but since the rest of the openobex package is not added, this results
in a "error: obex-check-device not found".
We do not need bluetooth stuff in the initrd anyway, so we remove the rules
file manually from the initrd until mkinitrd gets fixed.