Add support for persistence using a loop-mounted container file.
This adds an option to store persistent data in a container file instead of
a directory in the USB stick's Linux filesystem.
Theoretically, this would allow the Slackware Live files to be copied to
a VFAT partition on a stick without erasing it, as long as the stick is
made bootable and the Live kernel & initrd are added to the boot menu.

To create a USB stick from the Live ISO which uses a persistence file instead
of a persistence directory, run the 'iso2usb.sh' script with a new parameter:

  -P|--persistfile           Use a 'persistence' container file instead of
                             a directory (for use on FAT filesystem).

The following example creates a 750 MB LUKS-encrypted container file
'slhome.img' which will contain /home ; as well as a file '/persistence.img'
in the root of the USB's Linux partition which will be used to store the
Live OS' persistent data:

  ./iso2usb.sh -i slackware64-live-current.iso -o /dev/sdX -c 750M -P

NOTE:
  The persistence container file will be created as a 'sparse' file which
  starts as an empty file not using up any disk space, but is allowed to
  grow dynamically to consume a maximum value of 90% of the free disk space
  on the Linux partition of the USB stick.