Live init: add 'toram' boot parameter.
This allows the user to remove the bootable Live media afterwards, because
the complete OS will been copied to RAM and run from there when adding
the boot commandline parameter 'toram'.

This feature works nicely for the XFCE ISO since that one is fairly small.
Untested for the bigger ISOs; you need more patience while the squashfs
modules are being copied to RAM, and available RAM decreases a lot.
Realistically speaking, your computer would need 1 or 2 GB of RAM for
the XFCE variant, 2 or 4 GB RAM for the Mate variant, and 4 to 8 GB RAM
for the KDE4 and Plasma5 variants.