Add support for 32-bit UEFI boot (disabled by default).
This requires functionality in grub which is currently not enabled
in Slackware's grub package.
If you want a 32-bit Live ISO that boots on UEFI computers, you need
to recompile Slackware's grub with the patch for grub.SlackBuild which
you can find in the ./patches subdirectory.
Then you need to set the variable "EFI32" to "YES" in the
'make_slackware_live.sh' script and (re-)generate your 32-bit ISO.
The resulting 32-bit Live ISO will be UEFI-bootable.