Deal properly with the user enabling any optional/addon kernel modules.
When the user loads a non-system squashfs module (for instance from the
optional or addons subdirectories) and this contains kernel modules,
then the kernel must be made aware using a "depmod -a" call.

The previous implementation did this at the end of rc.local but
that was too late in the boot process: the kernel needs to be aware
of the new modules straight from the start, so that udev can enable
the hardware which is supported by these new kernel drivers on time.
So the depmod call was moved to the live init script (inside the initrd).