iso2usb.sh: added 'refresh' parameter to update an existing Live USB stick.
-r|--refresh =>
  Refresh an existing Live USB stick with new ISO content.

The refresh option "-r" can be used when you have a Live USB stick and you
want to update the liveslak files with new versions on an ISO image file
which you downloaded/created.

The USB stick will not be formatted in this case, and user modifications
like LUKS-encrypted homedirectory or persistence data will not be touched.
The "system" squashfs modules will be replaced with the versions on the ISO
file and any squashfs module not present in the system/ directory of the ISO
will be deleted.
The "optional" and "addons" directories are treated a bit differently:
The files in there that you added yourself will not be deleted when
the content of these directories is refreshed with the content from the ISO
(unless of course your added files have the same name as one of the files
in the ISO).
The content of the /boot and /EFI directories will both be replaced with
whatever is on the ISO image.