liveinit: allow modules to be organized in subdirectories.
The addons/ and optional/ directory will be searched for *.sxz modules
even if they live in subdirectories.  This allows you to organize your
modules instead of dumping them all into a single directory.

If a module fails to load (because it is corrupted for instance) liveinit
will mention this on screen and write the modulename to a file
'/mnt/live/modules/failed'.