Added package lists and repository info for the LXQT desktop environment.
Note that the 'alien.*' package list which is used for the PLASMA5 ISO
also adds the full LXQT desktop.  LXQT has a lot of dependencies (covered
in the 'lxqtdeps.lst) but almost all of these dependencies are already
part of PLASMA5 anyway.
The two 'lxqt.*' and 'lxqtdeps.*' package lists are meant for people
who do not want to create a PLASMA5 ISO (and therefore do not want to use
the 'alien.lst') but want an ISO focusing on LXQT without all the plasma5
or alien stuff.