(e)liloconfig need patching to make them work properly on Slackware Live.
The cause is the use of hard-coded '/mnt' instead of using $T_PX variable.
In Slackware's official installer, the computer's hard disk is always mounted
at /mnt so that there is no apparent issue.
Slackware Live's 'setup2hd' however, needs to mount the hard disk on /setup2hd
because /mnt is already used.