PLASMA5: add several new packages I forgot earlier.
In fact, I had not intended to replace Slackware's kdelibs4-based calligra
with the new KF5-based version.  But upgrading calligra allowed me to get
rid of the old kdelibs4-based kactivities, nepomuk-core and nepomuk-widgets
packages.
And krita has been made separate from calligra, so that is a new package
in my 'ktown' set altogether.  Kexi is absent (it was split out of calligra
just like krita) but if someone needs it, I will add it in the next update.
And finally, the package libkcddb4 is the older ldelibs4-based version
which is required by k3b.
The new KF5-based libkcddb is incompatible with k3b.