Make the 'make-grub.sh' script compatible with grub-2.02
In particular, it seems that including the 'ahci' grub module caused failure
on boot, because the actual Live USB stick would not be detected and the
local harddisk would be detected as (ahci0).
The script has also been enhanced so that the 'grub-mkimage' commandline will
not get fed any non-existing module, which would otherwise abort the command.