XFCE: rename xbase.lst to x_base.lst
The order in which the squashfs modules were created should be respected
when they are re-assembled to an overlay filesystem.
The assembly is done in  lexicographical order - therefore xbase.lst
needed to be renamed to x_base.lst.

Failing to adhere to the original assembly order causes weirdness.
In my case it meant that the complete directory /usr/share/applications
as contained in xapbase.lst was invisible in the resulting overlay, causing
near-empty menus in XFCE.

Also added several packages to x_base.lst that are now part of Slackware.