Add support for on-disk OS configuration file
The configuration file must be in the liveslak main directory ('/liveslak'
by default) and be called 'DISTRO_os.cfg' where DISTRO is the lower-case
name of the distro ('slackware' by default).

The following variables can be defined in the configuration file,
one per line, in the format VAR=value:
BLACKLIST, INIT, KEYMAP, LIVE_HOSTNAME, LOAD, LOCALE, LUKSVOL, NOLOAD,
RUNLEVEL, TWEAKS, TZ, XKB.

Since the configuration file is sourced by the live init script, it is OK
to have comments in it that are prefixed with '#'.