Slackware Live Edition: version 1.3.0
The fairly large version bump is warranted because this is the first
release to support zstd compression. The resulting ISO will boot to
a usable desktop environment almost twice as fast as previous releases
of liveslak. THe runtime behaviour will also be a lot snappier.
This should make a lot of people happy.

Note that I have to keep the XFCE ISO xz-compressed to keep its size
below 703 MB. The other ISOs will use zstd compression, adding roughly
10% to the ISO size but speeding up the Live OS with 60% or more.